refactor grin crate into separate modules
