+++ servers/src/grin/mod.rs
@@ -12,26 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Crate wrapping up the Grin binary and configuration file
+//! Grin P2P / API server
 
-#![deny(non_upper_case_globals)]
-#![deny(non_camel_case_types)]
-#![deny(non_snake_case)]
-#![deny(unused_mut)]
-#![warn(missing_docs)]
-
-extern crate serde;
-#[macro_use]
-extern crate serde_derive;
-extern crate toml;
-
-extern crate grin_servers as servers;
-extern crate grin_p2p as p2p;
-extern crate grin_pow as pow;
-extern crate grin_util as util;
-extern crate grin_wallet as wallet;
-
-pub mod config;
-pub mod types;
-
-pub use types::{ConfigError, ConfigMembers, GlobalConfig};
+pub mod dandelion_monitor;
+pub mod seed;
+pub mod server;
+pub mod sync;
