+++ servers/tests/dandelion.rs
@@ -20,7 +20,7 @@ extern crate grin_api as api;
 extern crate grin_chain as chain;
 extern crate grin_config as config;
 extern crate grin_core as core;
-extern crate grin_grin as grin;
+extern crate grin_servers as servers;
 extern crate grin_p2p as p2p;
 extern crate grin_pow as pow;
 extern crate grin_util as util;
